+++ packages/next-common/components/assets/context/assetHubTabsProvider.js
@@ -1,8 +1,5 @@
 import { createStateContext } from "react-use";
-import { useCallback, useEffect } from "react";
-import useMyAssets from "next-common/components/assets/useMyAssets";
-import useTransfersHistory from "next-common/utils/hooks/useTransfersHistory";
-import { defaultPageSize } from "next-common/utils/constants";
+import { useCallback } from "react";
 
 export const TABS = Object.freeze({
   assets: 1,
@@ -26,6 +23,8 @@ const [useTransfersHistoryContext, TransfersHistoryProvider] =
     total: 0,
   });
 
+export { useTransfersHistoryContext, useAssetsContext };
+
 export const useActiveTab = () => {
   const [activeTabId, setActiveTabId] = useActiveTabContext();
   return [activeTabId, setActiveTabId];
@@ -45,43 +44,6 @@ export const useTotalCounts = () => {
   );
 
   return [totalCounts, setTotalCount];
-};
-
-export const useAssets = () => {
-  const [assets, setAssets] = useAssetsContext();
-  const [, setTotalCount] = useTotalCounts();
-  const fetchedAssets = useMyAssets();
-
-  useEffect(() => {
-    if (fetchedAssets) {
-      setAssets(fetchedAssets);
-      setTotalCount("assets", fetchedAssets?.length || 0);
-    }
-  }, [fetchedAssets, setAssets, setTotalCount]);
-
-  return assets;
-};
-
-export const useQueryTransfersHistory = (page) => {
-  const [state, setState] = useTransfersHistoryContext();
-  const [, setTotalCount] = useTotalCounts();
-  const { value, total, loading, error } = useTransfersHistory(
-    page,
-    defaultPageSize,
-  );
-
-  useEffect(() => {
-    if (!loading && !error && value) {
-      setState({ list: value, total });
-      setTotalCount("transfers", total);
-    }
-  }, [loading, error, value, total, setState, setTotalCount]);
-
-  return {
-    list: state.list,
-    total: state.total,
-    loading,
-  };
 };
 
 export function AssetHubTabsProvider({ children }) {