+++ packages/next-common/utils/hooks/useTransfersHistory.js
@@ -1,5 +1,4 @@
 import { useState, useEffect } from "react";
-import useRealAddress from "next-common/utils/hooks/useRealAddress";
 import { useChain } from "next-common/context/chain";
 import { gql, request } from "graphql-request";
 
@@ -39,8 +38,7 @@ const STATESCAN_CHAIN_URL_MAP = Object.freeze({
   "polkadot-assethub": "https://statemint-gh-api.statescan.io/graphql",
 });
 
-export default function useTransfersHistory(page = 0, page_size = 25) {
-  const address = useRealAddress();
+export default function useTransfersHistory(address, page = 0, page_size = 25) {
   const chain = useChain();
 
   const [value, setValue] = useState([]);